Enhanced polycurve
