+++ Arrangement_on_surface_2/include/CGAL/Arr_geometry_traits/Polycurve_2.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2009,2010,2011 Tel-Aviv University (Israel).
+// Copyright (c) 2009,2010,2011,2015 Tel-Aviv University (Israel).
 // All rights reserved.
 //
 // This file is part of CGAL (www.cgal.org).
@@ -12,257 +12,494 @@
 // This file is provided AS IS with NO WARRANTY OF ANY KIND, INCLUDING THE
 // WARRANTY OF DESIGN, MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE.
 //
-// $URL$
-// $Id$
-// 
-//
-// Author(s)     : Efi Fogel <efif@post.tau.ac.il>
+// Author(s)     : Ron Wein  <wein@post.tau.ac.il>
+//                 Efi Fogel <efif@post.tau.ac.il>
+//                 Dror Atariah <dror.atariah@fu-berlin.de>
 
 #ifndef CGAL_ARR_POLYCURVE_2_H
 #define CGAL_ARR_POLYCURVE_2_H
 
 /*! \file
- * Header file for the polycurve classes used by the Arr_polycurve_traits_2
- * class.
+ * Header file for the polyline classes used by the
+ * Arr_polycurve_basic_traits_2, Arr_polycurve_traits_2, and
+ * Arr_polyline_traits_2 classes.
  */
 
 #include <list>
 #include <iostream>
 #include <vector>
 #include <iterator>
-
 #include <CGAL/Bbox_2.h>
 
 namespace CGAL {
 
+namespace internal {
 
 /*! \class
  * Representation of a polycurve.
  */
-template <class CurveTraits>
-class Polycurve_2
-{
+template <typename SubcurveType_2, typename PointType_2>
+class Polycurve_2 {
 public:
-  typedef CurveTraits                                   Curve_traits_2;
-  typedef typename Curve_traits_2::Point_2              Point_2;
-  typedef typename Curve_traits_2::X_monotone_curve_2   Curve_2;
-  typedef typename Curve_traits_2::Curve_2              Curve_2;
+  typedef SubcurveType_2                                Subcurve_type_2;
+  typedef PointType_2                                   Point_type_2;
 
-  typedef typename std::vector<Curve_2>::const_iterator          const_iterator;
-  typedef typename std::vector<Curve_2>::const_reverse_iterator  const_reverse_iterator;
-  
 protected:
-  // The segments that comprise the polycurve:
-  std::vector<Curve_2> curves;
+  // The subcurves that comprise the poyline:
+  typedef typename std::vector<Subcurve_type_2>         Subcurves_container;
+
+  Subcurves_container m_subcurves;
 
 public:
-  /*! Default constructor. */
-  Polycurve_2() {}
+  typedef typename Subcurves_container::size_type       Size;
+  typedef typename Subcurves_container::size_type       size_type;
 
-  /*!
-   * Constructor from a range of curves.
-   * \param begin An iterator pointing to the first curve in the range.
-   * \param end An iterator pointing after the last curve in the range.
-   * \pre the target of the i-th curve is identical to the source of the
-   * i+1-th curve for i = 0,1,...,n-1
+  /*! Construct default. */
+  Polycurve_2() : m_subcurves() {}
+
+  /*! Construct from a subcurve. */
+  Polycurve_2(const Subcurve_type_2& seg) : m_subcurves()
+  { m_subcurves.push_back(seg); }
+
+  /* Roadmap:
+   * - Currently construction from points is marked as deprecated.
+   *   For the sake of backwards compatibility, it is not removed yet.
+   * - In the next release, construction from points will be removed.
+   * - Then, the remaining construction from subcurves will become private
+   *   in this class as construction is provided by the traits.
+   *   Furthermore, the polycurve traits class will become a friend of
+   *   this class.
    */
-  template <class InputIterator>
-  Polycurve_2(InputIterator begin, InputIterator end)
+
+  /*! Construct from a range. The range can be either:
+   *
+   * For the sake of backwards compatibility we have to keep the possibility
+   * of construction from a range of points. Therefore, we have to test the
+   * type of the elements of the range again.
+   * Since ultimately the construction from points will be deprecated, the
+   * actual constructor that we implemented which handles a range of point
+   * is already deprecated and SHOULD NOT BE USED.
+   *
+   * If you want to construct a polycurve from a range of points, use the
+   * construction functors from the traits class.
+   *
+   * - Range of points, and the polycurve is defined by the order of the
+   *   points.
+   * - Range of linear object. The polycurve is the sequence of linear
+   *   objects.
+   * \param begin An iterator pointing to the first point in the range.
+   * \param end An iterator pointing after the last point in the range.
+   * \pre Depends on the range's content. See the implementations for
+   *      further details.
+   */
+  template <typename InputIterator>
+  Polycurve_2(InputIterator begin, InputIterator end) :
+    m_subcurves()
   {
-    std::copy(begin, end, std::back_inserter(curves));
+    typedef typename std::iterator_traits<InputIterator>::value_type VT;
+    typedef typename boost::is_same<VT, Point_type_2>::type Is_point;
+    construct_polycurve(begin, end, Is_point());
   }
 
-  /*! Append a curve to the polycurve.
-   * \param curve The new curve
-   * \pre the target of the current last curve must be identical to the
-   * source of curve
+  /*! Construct a polycurve from a range of subcurves.
+   * \param begin An iterator pointing to the first subcurve in the range.
+   * \param end An iterator pointing after the past-the-end subcurve
+   *        in the range.
+   * \pre The end of subcurve n should be the beginning of subcurve n+1.
    */
-  void push_back(const Curve & curve)
+  template <typename InputIterator>
+  void construct_polycurve(InputIterator begin, InputIterator end,
+                          boost::false_type)
+  { m_subcurves.assign(begin, end); }
+
+  /*! Construct a polycurve from a range of points.
+   * \param begin An iterator pointing to the first point in the range.
+   * \param end An iterator pointing after the last point in the range.
+   * \pre There are at least 2 points in the range.
+   *      In other cases, an empty polycurve will be created.
+   */
+  template <typename InputIterator>
+  CGAL_DEPRECATED void construct_polycurve(InputIterator begin,
+                                          InputIterator end,
+                                          boost::true_type)
   {
-    curves.push_back (curve);
+    // Check if there are no points in the range:
+    InputIterator ps = begin;
+    if (ps == end) return;
+
+    InputIterator pt = ps;
+    ++pt;
+
+    // The range contains only one point. A degenerated polycurve is
+    // constructed.
+    // With one degenerated subcurve, where source=target.
+    if (pt == end) {
+      m_subcurves.push_back(Subcurve_type_2(*ps, *ps));
+      return;
+    }
+
+    // Construct a subcurve from each to adjacent points.
+    // The container has to contain at least two points.
+    while (pt != end) {
+      m_subcurves.push_back(Subcurve_type_2(*ps, *pt));
+      ++ps;
+      ++pt;
+    }
   }
 
-  /*!
-   * Create a bounding-box for the polycurve.
+  /* Roadmap: Make this private in the next version (after the traits
+   *          becomes friendly...)
+   */
+
+  /*! Append a subcurve to the (x-monotone) polycurve.
+   * Warning: This is a risky function! Don't use it! Prefer the
+   *          provided implementation in the traits class.
+   * \param seg The new subcurve to be appended to the polycurve.
+   * \pre If the polycurve is not empty, seg source must be the
+   *      same as the target point of the last subcurve in the polycurve.
+   */
+  inline void push_back(const Subcurve_type_2& seg)
+  { this->m_subcurves.push_back(seg); }
+
+  /*! Append a subcurve to the (x-monotone) polycurve.
+   * Warning: This is a risky function! Don't use it! Prefer the
+   *          provided implementation in the traits class.
+   * \param seg The new subcurve to be appended to the polycurve.
+   * \pre If the polycurve is not empty, seg source must be the
+   *      same as the target point of the last subcurve in the polycurve.
+   */
+  inline void push_front(const Subcurve_type_2& seg)
+  { this->m_subcurves.insert(this->m_subcurves.begin(), seg); }
+
+  /*! Append a point to the polycurve.
+   * To properly implemented this function the traits class is needed,
+   * thus it is deprecated.
+   */
+  CGAL_DEPRECATED void push_back(const Point_type_2& p)
+  {
+    CGAL_assertion(!m_subcurves.empty());
+    Point_type_2 pt = p;
+    Point_type_2 ps = m_subcurves.back().target();
+    m_subcurves.push_back(Subcurve_type_2(ps, pt));
+  }
+
+  /*! Compute the bounding box of the polycurve.
    * \return The bounding-box.
    */
   Bbox_2 bbox() const
   {
-    // Compute the union of the bounding boxes of all curves.
-    unsigned int n = size();
+    // Compute the union of the bounding boxes of all subcurves.
+    size_type n = this->number_of_subcurves();
     Bbox_2 bbox;
-    unsigned int  i;
-
-    const_iterator ci = begin();
-    bbox = ci->bbox(*ci);
-    for (++ci; ci != end(); ++ci)
-      bbox = bbox + ci->bbox(*ci);
+    for (std::size_t i = 0; i < n; ++i)
+      bbox = (i > 0) ? (bbox + (*this)[i].bbox()) : (*this)[i].bbox();
     return bbox;
   }
 
-  /*! Obtain an iterator for the polycurve points. */
-  const_iterator begin() const { return curves.begin(); }
+  class const_iterator;
+  friend class const_iterator;
+  typedef std::reverse_iterator<const_iterator>
+  const_reverse_iterator;
 
-  /*! Obtain a past-the-end iterator for the polycurve points. */
-  const_iterator end() const { return curves.end(); }
+  /*! An iterator for the polycurve points. */
+  class CGAL_DEPRECATED const_iterator {
+  public:
+    // Type definitions:
+    typedef std::bidirectional_iterator_tag     iterator_category;
+    typedef Point_type_2                        value_type;
+    typedef std::ptrdiff_t                      difference_type;
+    typedef size_t                              size_type;
+    typedef const value_type&                   reference;
+    typedef const value_type*                   pointer;
 
-  /*! Obtain an reverse iterator for the polycurve points. */
-  const_reverse_iterator rbegin() const { return curves.rbegin(); }
+  private:
+    // The polycurve curve.
+    const Polycurve_2<Subcurve_type_2, Point_type_2>* m_cvP;
+    int m_num_pts;                            // Its number of points.
+    int m_index;                              // The current point.
+
+    /*! Private constructor.
+     * \param cv The scanned curve.
+     * \param index The index of the subcurve.
+     */
+    const_iterator(const Polycurve_2<Subcurve_type_2, Point_type_2>* cvP,
+                   int index) :
+      m_cvP(cvP),
+      m_index(index)
+      {
+        if (m_cvP == NULL)
+          m_num_pts = 0;
+        else
+          m_num_pts = (m_cvP->number_of_subcurves() == 0) ?
+            0 : (m_cvP->number_of_subcurves() + 1);
+      }
+
+  public:
+    /*! Default constructor. */
+    const_iterator() :
+      m_cvP(NULL),
+      m_num_pts(0),
+      m_index(-1)
+      {}
+
+    /*! Dereference operator.
+     * \return The current point.
+     */
+    const Point_type_2& operator*() const
+    {
+      CGAL_assertion(m_cvP != NULL);
+      CGAL_assertion((m_index >= 0) && (m_index < m_num_pts));
+
+      // First point is the source of the first subcurve.
+      // Else return the target of the(i-1)'st subcurve.
+      if (m_index == 0) return ((*m_cvP)[0]).source();
+      else return ((*m_cvP)[m_index - 1]).target();
+    }
+
+    /*! Arrow operator.
+     * \return A pointer to the current point.
+     */
+    const Point_type_2* operator->() const { return (&(this->operator* ())); }
+
+    /*! Increment operators. */
+    const_iterator& operator++()
+    {
+      if ((m_cvP != NULL) && (m_index < m_num_pts)) ++m_index;
+      return (*this);
+    }
+
+    const_iterator operator++(int)
+    {
+      const_iterator temp = *this;
+      if ((m_cvP != NULL) && (m_index < m_num_pts)) ++m_index;
+      return temp;
+    }
+
+    /*! Decrement operators. */
+    const_iterator& operator--()
+    {
+      if ((m_cvP != NULL) && (m_index >= 0)) --m_index;
+      return (*this);
+    }
+
+    const_iterator operator--(int)
+    {
+      const_iterator  temp = *this;
+      if ((m_cvP != NULL) && (m_index >= 0))
+        --m_index;
+      return temp;
+    }
+
+    /*! Equality operators. */
+    bool operator==(const const_iterator& it) const
+    { return ((m_cvP == it.m_cvP) && (m_index == it.m_index)); }
+
+    bool operator!=(const const_iterator& it) const
+    { return ((m_cvP != it.m_cvP) || (m_index != it.m_index)); }
+
+    friend class Polycurve_2<Subcurve_type_2, Point_type_2>;
+  };
+
+  /* ! Obtain an iterator for the polycurve points.*/
+  CGAL_DEPRECATED const_iterator begin() const
+  {
+    if (number_of_subcurves() == 0) return (const_iterator(NULL, -1));
+    else return (const_iterator(this, 0));
+  }
+
+  /*! Obtain a past-the-end iterator for the polycurve points.*/
+  CGAL_DEPRECATED const_iterator end() const
+  {
+    if (number_of_subcurves() == 0) return (const_iterator(NULL, -1));
+    else return (const_iterator(this, number_of_subcurves() + 1));
+  }
+
+  /*! Obtain a reverse iterator for the polycurve points. */
+  CGAL_DEPRECATED const_reverse_iterator rbegin() const
+  { return (const_reverse_iterator(end())); }
 
   /*! Obtain a reverse past-the-end iterator for the polycurve points. */
-  const_reverse_iterator rend() const { return curves.rend(); }
+  CGAL_DEPRECATED const_reverse_iterator rend() const
+  { return (const_reverse_iterator(begin())); }
 
-  /*!
-   * Obtain the number of curves that comprise the polycurve.
-   * \return The number of curves.
-   */
-  inline unsigned int size() const { return curves.size(); }
+  typedef typename Subcurves_container::iterator Subcurve_iterator;
+  typedef typename Subcurves_container::const_iterator
+                                                 Subcurve_const_iterator;
+  typedef typename std::reverse_iterator<Subcurve_const_iterator>
+                                                 Subcurve_const_reverse_iterator;
 
-  /*!
-   * Obtain the i'th segment of the polycurve.
-   * \param i The segment index(from 0 to size()-1).
-   * \return A const reference to the segment.
+  /*! Obtain an iterator for the polycurve subcurves. */
+  Subcurve_const_iterator begin_subcurves() const
+  { return m_subcurves.begin(); }
+
+  /*! Obtain a past-the-end iterator for the polycurve subcurves. */
+  Subcurve_const_iterator end_subcurves() const
+  { return m_subcurves.end(); }
+
+  /*! Obtain a reverse iterator for the polycurve subcurves. */
+  Subcurve_const_reverse_iterator rbegin_subcurves() const
+  { return (Subcurve_const_reverse_iterator(end_subcurves())); }
+
+  /*! Obtain a reverse past-the-end iterator for the polycurve points. */
+  Subcurve_const_reverse_iterator rend_subcurves() const
+  { return (Subcurve_const_reverse_iterator(begin_subcurves())); }
+
+  /*! Deprecated!
+   * Get the number of points contained in the polycurve.
+   * In general (for example if the polycurve is not bounded), then the
+   * number of vertices cannot be read-off from the number of subcurves, and
+   * the traits class is needed.
+   * \return The number of points.
    */
-  inline const Curve_2 & operator[](const unsigned int i) const
-  { return curves[i]; }
+  CGAL_DEPRECATED std::size_t points() const
+  { return (number_of_subcurves() == 0) ? 0 : number_of_subcurves() + 1; }
+
+  /*! Deprecated! Replaced by number_of_subcurves()
+   * Get the number of subcurves that comprise the poyline.
+   * \return The number of subcurves.
+   */
+  CGAL_DEPRECATED size_type size() const
+  { return m_subcurves.size(); }
+
+  /*! Obtain the number of subcurves that comprise the poyline.
+   * \return The number of subcurves.
+   */
+  size_type number_of_subcurves() const
+  { return m_subcurves.size(); }
+
+  /*! Obtain the ith subcurve of the polycurve.
+   * \param i The subcurve index(from 0 to size()-1).
+   * \return A const reference to the subcurve.
+   */
+  inline const Subcurve_type_2& operator[](const std::size_t i) const
+  {
+    CGAL_assertion(i < number_of_subcurves());
+    return (m_subcurves[i]);
+  }
 
   /*! Clear the polycurve. */
-  inline void clear() { curves.clear(); }
+  inline void clear() { m_subcurves.clear(); }
 };
 
 /*! \class
  * Representation of an x-monotone polycurve.
  * An x-monotone polycurve is always directed from left to right.
  */
-template <class CurveTraits>
-class X_monotone_polycurve_2 : public Polycurve_2<CurveTraits>
+template <typename SubcurveType_2, typename PointType_2>
+class X_monotone_polycurve_2 :
+    public Polycurve_2<SubcurveType_2, PointType_2>
 {
 public:
-  typedef CurveTraits                           Curve_traits_2;
-  typedef Polycurve_2<Curve_traits_2>           Base;
-  typedef typename Curve_traits_2::Point_2      Point_2;
-  typedef typename Curve_traits_2::Curve_2      Curve_2;
+  typedef SubcurveType_2                             Subcurve_type_2;
+  typedef PointType_2                                Point_type_2;
 
-  /*! Default constructor. */
+  typedef Polycurve_2<Subcurve_type_2, Point_type_2> Base;
+
+  /*! Construct default. */
   X_monotone_polycurve_2() : Base() {}
 
-  /*!
-   * Constructor from a range of points, defining the endpoints of the
-   * polycurve curves.
+  /*! Construct from a subcurve. */
+  X_monotone_polycurve_2(Subcurve_type_2 seg) : Base(seg) {}
+
+  /*! Construct from a range.
+   * Similar to the constructor of a general polycurve.
+   * Like in the case of general polycurve, for the sake of backwards
+   * compatibility we have to keep an implementation of construction
+   * from a range of points. DO NOT USE THIS CONSTRUCTION.
    */
-  template <class InputIterator>
+  template <typename InputIterator>
   X_monotone_polycurve_2(InputIterator begin, InputIterator end) :
     Base(begin, end)
   {
-    // Make sure the range of points contains at least two points.
-    Curve_traits_2 seg_traits;
-    InputIterator ps = begin;
-    CGAL_precondition(ps != end);
-    InputIterator pt = ps;
-    ++pt;
-    CGAL_precondition(pt != end);
-
-    CGAL_precondition_code(
-      typename Curve_traits_2::Compare_x_2 compare_x =
-        seg_traits.compare_x_2_object();
-      );
-    CGAL_precondition_code(
-      typename Curve_traits_2::Compare_xy_2 compare_xy =
-        seg_traits.compare_xy_2_object();
-      );
-    
-    
-    // Make sure there is no change of directions as we traverse the polycurve.
-    CGAL_precondition_code(
-      const Comparison_result cmp_x_res = compare_x(*ps, *pt);
-    );
-    const Comparison_result cmp_xy_res = compare_xy(*ps, *pt);
-    CGAL_precondition(cmp_xy_res != EQUAL);
-    ++ps; ++pt;
-    while (pt != end) {
-      CGAL_precondition (compare_xy(*ps, *pt) == cmp_xy_res);
-      CGAL_precondition (compare_x(*ps, *pt) == cmp_x_res);
-      ++ps; ++pt;
-    }
-
-    // Reverse the polycurve so it always directed from left to right.
-    if (cmp_xy_res == LARGER)
-      _reverse();
+    typedef typename std::iterator_traits<InputIterator>::value_type VT;
+    typedef typename boost::is_same<VT, Point_type_2>::type Is_point;
+    construct_x_monotone_polycurve(begin, end, Is_point());
   }
 
-  /*!
-   * Append a segment to the polycurve.
-   * \param seg The new segment to be appended to the polycurve.
-   * \pre If the polycurve is not empty, the segment source must be the
-   *      same as the target point of the last segment in the polycurve
-   *      (thus it must extend it to the right).
+  /*! Construct from a range of subcurves.
+   * This constructor is expected to be called only from the
+   * traits class, after the input was verified there.
+   * \pre The range of subcurves form an x-monotone polycurve.
    */
-  inline void push_back (const Curve_2& seg)
-  {
-    CGAL_precondition_code (Curve_traits_2   seg_tr);
-    CGAL_precondition_code (const unsigned int n = this->size());
-    CGAL_precondition (seg_tr.compare_xy_2_object() (seg.source(),
-						     seg.target()) == SMALLER);
-    CGAL_precondition (n == 0 ||
-		       seg_tr.equal_2_object() (this->curves[n - 1].target(),
-						seg.source()));
+  template <typename InputIterator>
+  void construct_x_monotone_polycurve(InputIterator, InputIterator,
+                                      boost::false_type)
+  {}
 
-    this->curves.push_back (seg);
-  }
-
-private:
-
-  /*! Reverse the polycurve. */
-  void _reverse()
-  {
-    typename Base::const_reverse_iterator  ps = this->rbegin();
-    typename Base::const_reverse_iterator  pt = ps;
-    ++pt;
-
-    std::vector<Curve_2>  rev_segs (this->size());
-    unsigned int            i = 0;
-
-    while (pt != this->rend())
-    {
-      rev_segs[i] = Curve_2 (*ps, *pt);
-      ++ps; ++pt;
-      i++;
-    }
-
-    this->curves = rev_segs;
-    return;
-  }
-
+  /*! Construct from a range of points, defining the endpoints of the
+   * polycurve subcurves.
+   */
+  template <typename InputIterator>
+  CGAL_DEPRECATED void construct_x_monotone_polycurve(InputIterator,
+                                                     InputIterator,
+                                                     boost::true_type)
+  {}
 };
 
 /*! Output operator for a polycurve. */
-template <class CurveTraits>
-std::ostream& operator<< (std::ostream & os,
-			  const Polycurve_2<CurveTraits>& cv)
+template <typename SubcurveType_2, typename PointType_2>
+std::ostream& operator<<(std::ostream & os,
+                         const Polycurve_2<SubcurveType_2, PointType_2>& cv)
 {
-  typename Polycurve_2<CurveTraits>::const_iterator  iter = cv.begin();
+  typedef SubcurveType_2                          Subcurve_type_2;
+  typedef PointType_2                            Point_type_2;
 
-  // Print the number of points:
-  os << cv.points();
+  typedef Polycurve_2<Subcurve_type_2, Point_type_2> Curve_2;
 
-  while (iter != cv.end())
-  {
-    os << "  " << *iter;
-    ++iter;
+  typename Curve_2::Subcurve_const_iterator iter = cv.begin_subcurves();
+  while (iter != cv.end_subcurves()) {
+    if (iter == cv.begin_subcurves()) {
+      os << " " << *iter;
+      ++iter;
+    }
+    else {
+      os << " <-> " << *iter;
+      ++iter;
+    }
   }
   return (os);
 }
 
 /*! Input operator for a polycurve. */
-template <class CurveTraits>
-std::istream & operator>> (std::istream & is,
-                           Polycurve_2<CurveTraits> & pl)
+template <typename SubcurveType_2, typename PointType_2>
+std::istream& operator>>(std::istream& is,
+                         Polycurve_2<SubcurveType_2, PointType_2>& pl)
 {
-  std::cerr << "Not implemented yet!" << std::endl;
-  return is;
+  typedef SubcurveType_2                                Subcurve_type_2;
+  typedef PointType_2                                   Point_type_2;
+
+  typedef Polycurve_2<Subcurve_type_2, Point_type_2>    Curve_2;
+
+  // Read the number of input points.
+  unsigned int n_pts;
+  is >> n_pts;
+  CGAL_precondition_msg(n_pts > 1, "Input must contain at least two points");
+
+  // Read m_num_pts points to a list.
+  Point_type_2 p;
+  std::list<Point_type_2> pts;
+  for (unsigned int i = 0; i < n_pts; ++i) {
+    is >> p;
+    pts.push_back(p);
+  }
+
+  std::list<Subcurve_type_2> subcurves;
+  typename std::list<Point_type_2>::iterator curr = pts.begin();
+  typename std::list<Point_type_2>::iterator next = pts.begin();
+  ++next;
+  while (next != pts.end()) {
+    subcurves.push_back(Subcurve_type_2(*curr, *next));
+    ++curr;
+    ++next;
+  }
+
+  // Create the polycurve curve.
+  pl = Curve_2(subcurves.begin(), subcurves.end());
+
+  return (is);
 }
 
-
+} // namespace polycurve
 } //namespace CGAL
 
 #endif
